refactor: keep helpers local
