+++ src/temp-cleanup.js
@@ -1,13 +1,13 @@
 import fs from 'node:fs';
 import path from 'node:path';
 import { fileURLToPath } from 'node:url';
-import { parseArgs, tempStackPath } from './utils.js';
+import { parseArgs } from './utils.js';
 
 export async function main(argv = process.argv.slice(2)) {
 	const args = parseArgs(argv, { required: ['app-path', 'pr-number'] });
 	const appPath = args['app-path'];
 	const prNumber = args['pr-number'];
-	const tempPath = tempStackPath(appPath, prNumber);
+	const tempPath = `${appPath}-pr-${prNumber}`;
 
 	if (!fs.existsSync(tempPath)) {
 		console.log(`Temp stack ${tempPath} does not exist, nothing to clean up`);
