+++ src/utils.js
@@ -1,9 +1,3 @@
-import fs from 'node:fs';
-
-export function isObject(value) {
-	return value !== null && typeof value === 'object' && !Array.isArray(value);
-}
-
 export function parseArgs(argv, { required = [] } = {}) {
 	const result = {};
 	for (let i = 0; i < argv.length; i++) {
@@ -19,87 +13,4 @@ export function parseArgs(argv, { required = [] } = {}) {
 		}
 	}
 	return result;
-}
-
-export function appendGithubOutput(name, value) {
-	const outputFile = process.env.GITHUB_OUTPUT;
-	if (!outputFile) return;
-
-	fs.appendFileSync(outputFile, `${name}=${value}\n`);
-}
-
-export function tempStackPath(appPath, prNumber) {
-	return `${appPath}-pr-${prNumber}`;
-}
-
-// Collect all Host() values from traefik labels
-export function collectHosts(services) {
-	const allHosts = [];
-	for (const [, service] of Object.entries(services)) {
-		if (!service.labels) continue;
-		for (const label of service.labels) {
-			for (const match of label.matchAll(/Host\(`([^`]+)`\)/g)) {
-				if (!allHosts.includes(match[1])) allHosts.push(match[1]);
-			}
-		}
-	}
-	return allHosts;
-}
-
-// Pick best host for domain extraction
-// Prefers non-www subdomain hosts (3+ parts like x.jaw.dev) over bare domains (closepowerlifting.com)
-export function detectHost(services) {
-	const allHosts = collectHosts(services);
-	return allHosts.find((h) => h.split('.').length >= 3 && !h.startsWith('www.')) || allHosts.find((h) => h.split('.').length >= 3) || allHosts[0] || null;
-}
-
-export function dependsOnServices(service) {
-	const dependsOn = service?.depends_on;
-	if (Array.isArray(dependsOn)) {
-		return dependsOn.filter((name) => typeof name === 'string');
-	}
-	if (dependsOn && typeof dependsOn === 'object') {
-		return Object.keys(dependsOn);
-	}
-	return [];
-}
-
-export function dependentServices(services, rootService) {
-	if (!services || !Object.hasOwn(services, rootService)) {
-		return new Set();
-	}
-
-	const selected = new Set([rootService]);
-	const pending = [rootService];
-
-	while (pending.length > 0) {
-		const serviceName = pending.pop();
-		for (const dependency of dependsOnServices(services[serviceName])) {
-			if (!Object.hasOwn(services, dependency) || selected.has(dependency)) {
-				continue;
-			}
-			selected.add(dependency);
-			pending.push(dependency);
-		}
-	}
-
-	return selected;
-}
-
-export function createGitHubApi(token, repo) {
-	const apiBase = `https://api.github.com/repos/${repo}`;
-	const headers = {
-		Authorization: `Bearer ${token}`,
-		Accept: 'application/vnd.github+json',
-		'Content-Type': 'application/json',
-		'X-GitHub-Api-Version': '2022-11-28',
-	};
-
-	return async function githubApi(path, options = {}) {
-		const res = await fetch(`${apiBase}${path}`, { headers, ...options });
-		if (!res.ok) {
-			throw new Error(`GitHub API ${options.method || 'GET'} ${path}: ${res.status} ${await res.text()}`);
-		}
-		return res;
-	};
 }
